backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.0
langchain-aws==0.1.0
langchain-core==0.1.23
langgraph==0.0.20
neo4j==5.16.0
tree-sitter
tree-sitter-python
e2b==0.15.0
gitpython==3.1.41
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
boto3
python-multipart==0.0.6
google-genai
